feat(http): 支持HEAD方法并更新服务器配置
添加对HTTP HEAD方法的支持，处理方式与GET相同但不返回响应体
更新服务器默认端口为8083
修复状态码枚举类型转换问题

+++ src/response/http_response.cc
@@ -224,7 +224,8 @@ std::string HttpResponse::toString() const {
     std::string result;
 
     // 1. 状态行：HTTP/1.1 <状态码> <状态消息>
-    result = "HTTP/1.1 " + std::to_string(m_statusCode) + " " + m_statusMessage + "\r\n";
+    // 将枚举类型转换为整数后再转换为字符串
+    result = "HTTP/1.1 " + std::to_string(static_cast<int>(m_statusCode)) + " " + m_statusMessage + "\r\n";
 
     // 2. Content-Type头部
     if (!m_contentType.empty()) {
